backend/requirements.txt
Flask
flask-cors
gtts
google-generativeai
python-dotenv
